+++ game-pragmata/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - None Planned
 
-## [0.1.0] - 2026-04-24
+## [0.1.1] - 2026-04-23
+
+- Changed: REFramework now downloads from Nexus Mods page
+
+## [0.1.0] - 2026-04-17
 
 - Initial Release.