+++ game-tekken8/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - None
 
+## [0.1.1] - 2026-04-23
+
+- Fixed: Added better game image
+
 ## [0.1.0] - 2026-04-19
 
 - Initial release